code separated by folders
