+++ Agendaly/.idea/misc.xml
@@ -13,6 +13,7 @@
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/activity_horario.xml" value="0.25" />
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/activity_info_fragment.xml" value="0.2" />
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/activity_main.xml" value="0.1793478260869565" />
+        <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/authtest_activity.xml" value="0.17798913043478262" />
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/borrar_clases.xml" value="0.264" />
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/calendar_cell.xml" value="0.25" />
         <entry key="..\:/Users/laura/OneDrive/Escritorio/TT/Agendaly/app/src/main/res/layout/formulario_clases.xml" value="0.2" />
