+++ Agendaly/app/build.gradle
@@ -61,6 +61,6 @@ dependencies {
     // optional - Test helpers
     testImplementation "androidx.room:room-testing:$room_version"
     // optional - Paging 3 Integration
-    implementation "androidx.room:room-paging:2.4.0-beta01"
+    implementation "androidx.room:room-paging:2.4.0-rc01"
 
 }
