+++ Agendaly/app/src/main/java/es/udc/psi/agendaly/TimeTable/CalendarFragment.java
@@ -21,7 +21,7 @@ import devs.mulham.horizontalcalendar.HorizontalCalendar;
 import devs.mulham.horizontalcalendar.model.CalendarEvent;
 import devs.mulham.horizontalcalendar.utils.CalendarEventsPredicate;
 import devs.mulham.horizontalcalendar.utils.HorizontalCalendarListener;
-import es.udc.psi.agendaly.CalendarActivity;
+import es.udc.psi.agendaly.Calendar.CalendarActivity;
 import es.udc.psi.agendaly.R;
