+++ Agendaly/app/src/main/res/layout/activity_calendar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".CalendarActivity">
+    tools:context=".Calendar.CalendarActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
